Fix frequent reconnections when timers under/overshoot
The ping timer is set exactly on the interval. Sometimes the timer runs
a bit before the ping time so the ping misses. If the next timer
overshoots, the ping is not sent and the connection is restarted
(because the last message time is > ping_time*2). Scheduling the timer
more frequently than the actual ping time fixes these issues as the
timer is guaranteed to ping at least once during the ping_time*2
interval.