Support rack requests in Slack::Events::Request
The `#headers` method on a request is implemented by Rails, so in
order for the event request class to support other rack based servers
we can instead look for the headers via `#get_header` or `#env`.

Since Rail's requests are an extension of rack requests this will
also work in a Rails app.