Raise rescuable errors
This commit replaces all instances of `throw` with `raise`.

While I was using this library, I noticed that error handling didn't
work as expected.

Here's an example of what I tried. Looking at the code, I expected this
to print `ArgumentError`, but it printed `UncaughtThrowError`.

    client = Slack::Web::Client.new(token: ENV["TOKEN"])
    begin
      client.chat_postMessage
    rescue => e
      puts e.class
    end
    # UncaughtThrowError

According to Ruby's [tutorial on Exceptions][1], `raise` and `rescue`
should be used for exceptions, while `throw` and `catch` could be used
for control flow.

[1]: https://ruby-doc.com/docs/ProgrammingRuby/html/tut_exceptions.html